Follow up on changes to internal cards, stats
Always associate the product received by remote products such as LSSS.
This requires redesigning the Service and UI to associate with a given product.
Whereas before, the product was only associated on push, the product is now associated on create/edit.

The "Add an Internal Request" buttons needs a function instead of directly calling open modal.
This function can then explicitly refresh the products listing so that the list is available.