Partially revert refactor of projects to products
In the case of Remote Products, these are actually Remote Projects and should not have been refactored from Projects to Products.

In the case of `refreshRemoteProducts` method in the `productsService`, this is fetching remote projects associated for the given product.
Rename the method to be more accurate, now calling it `refreshRemoteProjectsForProduct`.
The remote projects in the products service exist as a cache for remote products specifically associated with a given product.
These remote projects, therefore, do not relate to the remote projects service.

In the case of `products/remote-projects/:productId` from `RemoteProjects`, the path is/shall-be moved to `projects/remote/by-product/:productId`.
Having the remote projects being loaded from within the products control and path is not a consistent design.
Selecting the remote projects is or shall be moved into the remote projects controller under the appropriate remote projects path in the Service.