Issue 70: Can Create invalid Remote Project Managers, breaking system.
This expands upon PR #73 further removing the basic authentication for VersionOne as well.

With these changes, it makes sense to remove the settings and simplify the model design.
This better allows for validation using the existing weaver validation functionality.
The most notable downside of this particular implementation is the database changes.

The URL is now a validated and required part of the ManagementService.

The token is now a validated part of the ManagementService, but is not explicitly required via the DB.
Both Github and VersionOne require token, so the default behavior is to require the token.

see: #70
see: #73